+++ src/dropdowns/Dropdown.jsx
@@ -13,6 +13,7 @@
  * 2024/10/11  ITA   1.04      Reduce the width of the text box so that it appears side by side with the drop-down on even smaller screens.
  * 2024/10/28  ITA   1.05      Improve the responsiveness of the Dropdown.
  * 2025/12/18  ITA   1.06      Performed further tweaks and tests, preparing this component for npm publishing.
+ * 2025/12/26  ITA   1.07      Changed the arrow symbols to + and - for better visibility across different platforms.
  */
 import PropTypes from 'prop-types';
 import { useState, useEffect } from 'react';
@@ -117,7 +118,7 @@ export function Dropdown({label, // label with which to describe the dropdown.
                         disabled={isDisabled} placeholder='Type to search' value={searchText} />
                 
                 <div className='dropdown-js-arrow-container dropdown-js-padding dropdown-js-rounded' onClick={e=> toggleShowList(e)}>
-                    <span className='dropdown-js-arrow dropdown-js-padding'>{!showItems? "\u25BE" : "\u25B4"}</span>
+                    <span className='dropdown-js-arrow dropdown-js-padding'>{!showItems? "+" : "-"}</span>
                 </div>
             </div>