+++ src/dropdowns/MultiSelectionDropdown.jsx
@@ -10,7 +10,8 @@
  *                              Adjust width and add borders.
  *                              Import context directly. Variable names moved to VarNames object.
  * 2024/10/28  ITA    1.03      Improve the responsiveness of the dropdown.
-*  2025/12/18  ITA    1.06      Performed further tweaks and tests, to prepare this component for npm publishing.
+ * 2025/12/18  ITA    1.06      Performed further tweaks and tests, to prepare this component for npm publishing.
+ * 2025/12/26  ITA    1.07      Changed the arrow symbols to + and - for better cross-platform rendering.
  */
 import PropTypes from 'prop-types';
 import { useState, useEffect } from 'react';
@@ -22,7 +23,7 @@ export function MultiSelectionDropdown({
                     collectionName,
                     isDisabled = false,
                     onItemsSelected = null,
-                    dropdownStyle,
+                    dropdownStyle, // Styling object with fields {color, backgroundColor, borderColor (optional), fontSize}.
                     buttonStyle // Styling for the DONE button
                 })
 {
@@ -157,7 +158,7 @@ export function MultiSelectionDropdown({
                 </div>
                 
                 <div className='dropdown-js-arrow-container dropdown-js-padding dropdown-js-rounded' onClick={e=> toggleShowList(e)}>
-                    <span className='dropdown-js-arrow dropdown-js-padding'>{!showItems? "\u25BE" : "\u25B4"}</span>
+                    <span className='dropdown-js-arrow dropdown-js-padding'>{!showItems? "+" : "-"}</span>
                 </div>
             </div>